+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Polyverse Boost ReST API (Backend)
 - Project Status reports if the Project was updated since the last synchronization to AI Servers
 - During Project Grooming Pass - if Project Data is out of date (e.g. Repo modified since last analysis synchronization), then re-initialize all Resource Generators
 - Simplify all generators to support 'Reset' re-initialization stage and start with a common StaticDefault stage
+- Ensure DELETE of /api/user_project/{org}/{project}/data_references also removes any OpenAI files associated with the project
 
 ### Bug Fixes
 - Fix error handling for missing repo param when retrieving a file from GitHub